Use Dev Services for Infinispan Client IT
