+++ integration-tests/infinispan-client/src/main/java/io/quarkus/it/infinispan/client/TestServlet.java
@@ -1,20 +1,12 @@
 package io.quarkus.it.infinispan.client;
 
 import java.math.BigDecimal;
-import java.time.YearMonth;
-import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
-import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.CompletionStage;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.TimeoutException;
 import java.util.stream.Collectors;
 
-import javax.enterprise.event.Observes;
 import javax.inject.Inject;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.GET;
@@ -27,13 +19,6 @@ import javax.ws.rs.core.Response;
 
 import org.infinispan.client.hotrod.RemoteCache;
 import org.infinispan.client.hotrod.Search;
-import org.infinispan.client.hotrod.annotation.ClientCacheEntryCreated;
-import org.infinispan.client.hotrod.annotation.ClientCacheEntryModified;
-import org.infinispan.client.hotrod.annotation.ClientCacheEntryRemoved;
-import org.infinispan.client.hotrod.annotation.ClientListener;
-import org.infinispan.client.hotrod.event.ClientCacheEntryCreatedEvent;
-import org.infinispan.client.hotrod.event.ClientCacheEntryModifiedEvent;
-import org.infinispan.client.hotrod.event.ClientCacheEntryRemovedEvent;
 import org.infinispan.client.hotrod.jmx.RemoteCacheClientStatisticsMXBean;
 import org.infinispan.client.hotrod.logging.Log;
 import org.infinispan.client.hotrod.logging.LogFactory;
@@ -41,20 +26,20 @@ import org.infinispan.counter.api.CounterConfiguration;
 import org.infinispan.counter.api.CounterManager;
 import org.infinispan.counter.api.CounterType;
 import org.infinispan.counter.api.StrongCounter;
-import org.infinispan.query.api.continuous.ContinuousQuery;
-import org.infinispan.query.api.continuous.ContinuousQueryListener;
 import org.infinispan.query.dsl.Query;
 import org.infinispan.query.dsl.QueryFactory;
 
 import io.quarkus.infinispan.client.Remote;
-import io.quarkus.runtime.StartupEvent;
 
 @Path("/test")
 public class TestServlet {
     private static final Log log = LogFactory.getLog(TestServlet.class);
 
     @Inject
-    @Remote("default")
+    CacheSetup cacheSetup;
+
+    @Inject
+    @Remote(CacheSetup.DEFAULT_CACHE)
     RemoteCache<String, Book> cache;
 
     @Inject
@@ -62,106 +47,16 @@ public class TestServlet {
     RemoteCache<String, Book> boolsOld;
 
     @Inject
-    @Remote("magazine")
+    @Remote(CacheSetup.MAGAZINE_CACHE)
     RemoteCache<String, Magazine> magazineCache;
 
     @Inject
     CounterManager counterManager;
 
-    CountDownLatch waitUntilStarted = new CountDownLatch(1);
-
-    final Map<String, Book> matches = new ConcurrentHashMap<>();
-
-    void onStart(@Observes StartupEvent ev) {
-        log.info("Servlet has started");
-
-        cache.addClientListener(new EventPrintListener());
-
-        log.info("Added client listener");
-
-        ContinuousQuery<String, Book> continuousQuery = Search.getContinuousQuery(cache);
-
-        QueryFactory queryFactory = Search.getQueryFactory(cache);
-        Query query = queryFactory.create("from book_sample.Book where publicationYear > 2011");
-
-        ContinuousQueryListener<String, Book> listener = new ContinuousQueryListener<String, Book>() {
-            @Override
-            public void resultJoining(String key, Book value) {
-                log.warn("Adding key: " + key + " for book: " + value);
-                matches.put(key, value);
-            }
-
-            @Override
-            public void resultLeaving(String key) {
-                log.warn("Removing key: " + key);
-                matches.remove(key);
-            }
-
-            @Override
-            public void resultUpdated(String key, Book value) {
-                log.warn("Entry updated: " + key);
-            }
-        };
-
-        continuousQuery.addContinuousQueryListener(query, listener);
-
-        log.info("Added continuous query listener");
-
-        cache.put("book1", new Book("Game of Thrones", "Lots of people perish", 2010,
-                Collections.singleton(new Author("George", "Martin")), Type.FANTASY, new BigDecimal("23.99")));
-        cache.put("book2", new Book("Game of Thrones Path 2", "They win?", 2023,
-                Collections.singleton(new Author("Son", "Martin")), Type.FANTASY, new BigDecimal("54.99")));
-
-        magazineCache.put("first-mad", new Magazine("MAD", YearMonth.of(1952, 10),
-                Collections.singletonList("Blob named Melvin")));
-        magazineCache.put("first-time", new Magazine("TIME", YearMonth.of(1923, 3),
-                Arrays.asList("First helicopter", "Change in divorce law", "Adam's Rib movie released",
-                        "German Reparation Payments")));
-        magazineCache.put("popular-time", new Magazine("TIME", YearMonth.of(1997, 4),
-                Arrays.asList("Yep, I'm gay", "Backlash against HMOS", "False Hope on Breast Cancer?")));
-
-        log.info("Inserted values");
-
-        waitUntilStarted.countDown();
-    }
-
-    @ClientListener
-    static class EventPrintListener {
-
-        @ClientCacheEntryCreated
-        public void handleCreatedEvent(ClientCacheEntryCreatedEvent e) {
-            log.warn("Someone has created an entry: " + e);
-        }
-
-        @ClientCacheEntryModified
-        public void handleModifiedEvent(ClientCacheEntryModifiedEvent e) {
-            log.warn("Someone has modified an entry: " + e);
-        }
-
-        @ClientCacheEntryRemoved
-        public void handleRemovedEvent(ClientCacheEntryRemovedEvent e) {
-            log.warn("Someone has removed an entry: " + e);
-        }
-
-    }
-
-    /**
-     * This is needed because start notification is done async - you can receive requests while running start
-     */
-    private void ensureStart() {
-        try {
-            if (!waitUntilStarted.await(10, TimeUnit.SECONDS)) {
-                throw new RuntimeException(new TimeoutException());
-            }
-        } catch (InterruptedException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
     @GET
     @Produces(MediaType.TEXT_PLAIN)
     public List<String> getIDs() {
-        ensureStart();
+        cacheSetup.ensureStarted();
         log.info("Retrieving all IDs");
         return cache.keySet().stream().sorted().collect(Collectors.toList());
     }
@@ -170,7 +65,7 @@ public class TestServlet {
     @GET
     @Produces(MediaType.TEXT_PLAIN)
     public String getCachedValue(@PathParam("id") String id) {
-        ensureStart();
+        cacheSetup.ensureStarted();
         Book book = cache.get(id);
         return book != null ? book.getTitle() : "NULL";
     }
@@ -179,7 +74,7 @@ public class TestServlet {
     @GET
     @Produces(MediaType.TEXT_PLAIN)
     public String queryAuthorSurname(@PathParam("id") String name) {
-        ensureStart();
+        cacheSetup.ensureStarted();
         QueryFactory queryFactory = Search.getQueryFactory(cache);
         Query query = queryFactory.from(Book.class)
                 .having("authors.name").like("%" + name + "%")
@@ -201,7 +96,7 @@ public class TestServlet {
     @GET
     @Produces(MediaType.TEXT_PLAIN)
     public String ickleQueryAuthorSurname(@PathParam("id") String name) {
-        ensureStart();
+        cacheSetup.ensureStarted();
         QueryFactory queryFactory = Search.getQueryFactory(cache);
         Query query = queryFactory.create("from book_sample.Book b where b.authors.name like '%" + name + "%'");
         List<Book> list = query.execute().list();
@@ -220,7 +115,7 @@ public class TestServlet {
     @GET
     @Produces(MediaType.TEXT_PLAIN)
     public CompletionStage<Long> incrementCounter(@PathParam("id") String id) {
-        ensureStart();
+        cacheSetup.ensureStarted();
         CounterConfiguration configuration = counterManager.getConfiguration(id);
         if (configuration == null) {
             configuration = CounterConfiguration.builder(CounterType.BOUNDED_STRONG).build();
@@ -234,8 +129,8 @@ public class TestServlet {
     @GET
     @Produces(MediaType.TEXT_PLAIN)
     public String continuousQuery() {
-        ensureStart();
-        return matches.values().stream()
+        cacheSetup.ensureStarted();
+        return cacheSetup.getMatches().values().stream()
                 .mapToInt(Book::getPublicationYear)
                 .mapToObj(Integer::toString)
                 .collect(Collectors.joining(","));
@@ -245,7 +140,7 @@ public class TestServlet {
     @GET
     @Produces(MediaType.TEXT_PLAIN)
     public String nearCache() {
-        ensureStart();
+        cacheSetup.ensureStarted();
         RemoteCacheClientStatisticsMXBean stats = cache.clientStatistics();
         long nearCacheMisses = stats.getNearCacheMisses();
         long nearCacheHits = stats.getNearCacheHits();
@@ -315,7 +210,7 @@ public class TestServlet {
     @PUT
     @Consumes(MediaType.TEXT_PLAIN)
     public Response createItem(String value, @PathParam("id") String id) {
-        ensureStart();
+        cacheSetup.ensureStarted();
         Book book = new Book(id, value, 2019, Collections.emptySet(), Type.PROGRAMMING, new BigDecimal("9.99"));
         Book previous = cache.putIfAbsent(id, book);
         if (previous == null) {
@@ -332,7 +227,7 @@ public class TestServlet {
     @Path("magazinequery/{id}")
     @GET
     public String magazineQuery(@PathParam("id") String name) {
-        ensureStart();
+        cacheSetup.ensureStarted();
         QueryFactory queryFactory = Search.getQueryFactory(magazineCache);
         Query query = queryFactory.create("from magazine_sample.Magazine m where m.name like '%" + name + "%'");
         List<Magazine> list = query.list();
