--- integration-tests/infinispan-client/src/main/resources/application.properties
@@ -1,8 +1,0 @@
-quarkus.infinispan-client.server-list=localhost:11232
-quarkus.infinispan-client.use-auth=false
-quarkus.infinispan-client.trust-store=src/main/resources/server.p12
-quarkus.infinispan-client.trust-store-password=changeit
-quarkus.infinispan-client.trust-store-type=PKCS12
-
-# quarkus.log.level=DEBUG
-# quarkus.log.console.level=DEBUG
