+++ integration-tests/infinispan-client/src/test/java/io/quarkus/it/infinispan/client/InfinispanClientFunctionalityTest.java
@@ -5,15 +5,10 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import org.junit.jupiter.api.Test;
 
-import io.quarkus.test.common.QuarkusTestResource;
 import io.quarkus.test.junit.QuarkusTest;
 import io.restassured.RestAssured;
 
-/**
- * @author William Burns
- */
 @QuarkusTest
-@QuarkusTestResource(InfinispanServerTestResource.class)
 public class InfinispanClientFunctionalityTest {
 
     @Test
